MGMT-9453: moving part of validation logic to pkg to use in agent
