+++ internal/dns/dns.go
@@ -8,11 +8,11 @@ import (
 
 	"github.com/danielerez/go-dns-client/pkg/dnsproviders"
 	"github.com/go-openapi/swag"
-	"github.com/openshift/assisted-service/internal/cluster/validations"
 	"github.com/openshift/assisted-service/internal/common"
 	"github.com/openshift/assisted-service/internal/network"
 	"github.com/openshift/assisted-service/models"
 	logutil "github.com/openshift/assisted-service/pkg/log"
+	"github.com/openshift/assisted-service/pkg/validations"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 )
@@ -199,8 +199,9 @@ func (h *handler) GetDNSDomain(clusterName, baseDNSDomainName string) (*DNSDomai
 // up to 63 bytes. A single char may occupy more than one byte in Internationalized Domain Names (IDNs).
 func (h *handler) ValidateDNSName(clusterName, baseDNSDomainName string) error {
 	appsDomainNameSuffix := fmt.Sprintf(appsDomainNameFormat, clusterName, baseDNSDomainName)
-	if err := validations.ValidateDomainNameFormat(baseDNSDomainName); err != nil {
-		return err
+	apiErrorCode, err := validations.ValidateDomainNameFormat(baseDNSDomainName)
+	if err != nil {
+		return common.NewApiError(apiErrorCode, err)
 	}
 	if len(appsDomainNameSuffix) > dnsDomainPrefixMaxLen {
 		return errors.Errorf("Combination of cluster name and base DNS domain too long")
